# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ERP Venezuela API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  DEVELOPER: "DEVELOPER",
  DUENO: "DUENO",
  CAJERA: "CAJERA",
} as const;

export interface User {
  id: number;
  username: string;
  name: string;
  role: UserRole;
  isActive: boolean;
  createdAt: string;
}

export interface LoginResponse {
  user: User;
  token: string;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  DEVELOPER: "DEVELOPER",
  DUENO: "DUENO",
  CAJERA: "CAJERA",
} as const;

export interface CreateUserRequest {
  username: string;
  name: string;
  password: string;
  role: CreateUserRequestRole;
}

export interface UpdateUserRequest {
  name?: string;
  isActive?: boolean;
  password?: string;
}

export interface Product {
  id: number;
  name: string;
  category: string;
  costUsd: number;
  priceUsd: number;
  stock: number;
  minStock: number;
  createdAt: string;
}

export interface CreateProductRequest {
  name: string;
  category: string;
  costUsd: number;
  priceUsd: number;
  stock: number;
  minStock: number;
}

export type SalePaymentMethod =
  (typeof SalePaymentMethod)[keyof typeof SalePaymentMethod];

export const SalePaymentMethod = {
  EFECTIVO_USD: "EFECTIVO_USD",
  PAGO_MOVIL: "PAGO_MOVIL",
  PUNTO_VENTA: "PUNTO_VENTA",
} as const;

export interface SaleItem {
  productId: number;
  productName: string;
  quantity: number;
  priceUsd: number;
  costUsd: number;
}

export interface Sale {
  id: number;
  cashierId: number;
  cashierName: string;
  totalUsd: number;
  totalBs: number;
  bcvRate: number;
  paymentMethod: SalePaymentMethod;
  items: SaleItem[];
  createdAt: string;
}

export type CreateSaleRequestPaymentMethod =
  (typeof CreateSaleRequestPaymentMethod)[keyof typeof CreateSaleRequestPaymentMethod];

export const CreateSaleRequestPaymentMethod = {
  EFECTIVO_USD: "EFECTIVO_USD",
  PAGO_MOVIL: "PAGO_MOVIL",
  PUNTO_VENTA: "PUNTO_VENTA",
} as const;

export interface CreateSaleItemRequest {
  productId: number;
  quantity: number;
}

export interface CreateSaleRequest {
  paymentMethod: CreateSaleRequestPaymentMethod;
  items: CreateSaleItemRequest[];
}

export interface TopProduct {
  productId: number;
  productName: string;
  totalQuantity: number;
  totalRevenue: number;
}

export interface DailySale {
  date: string;
  totalUsd: number;
  totalBs: number;
}

export interface SalesSummary {
  totalSalesUsd: number;
  totalSalesBs: number;
  totalCostUsd: number;
  netProfitUsd: number;
  salesCount: number;
  topProducts: TopProduct[];
  dailySales: DailySale[];
}

export type BcvRateSource = (typeof BcvRateSource)[keyof typeof BcvRateSource];

export const BcvRateSource = {
  SCRAPER: "SCRAPER",
  MANUAL: "MANUAL",
} as const;

export interface BcvRate {
  rate: number;
  source: BcvRateSource;
  updatedAt: string;
}

export interface SetBcvRateRequest {
  rate: number;
}

export type ListSalesParams = {
  /**
   * Start date (ISO)
   */
  from?: string;
  /**
   * End date (ISO)
   */
  to?: string;
};

export type GetSalesSummaryParams = {
  from?: string;
  to?: string;
};