# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ERP Venezuela API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login with username and password
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    username: zod.string(),
    name: zod.string(),
    role: zod.enum(["DEVELOPER", "DUENO", "CAJERA"]),
    isActive: zod.boolean(),
    createdAt: zod.string(),
  }),
  token: zod.string(),
});

/**
 * @summary Logout current user
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Get current authenticated user
 */
export const GetCurrentUserResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  name: zod.string(),
  role: zod.enum(["DEVELOPER", "DUENO", "CAJERA"]),
  isActive: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary List all users (DEVELOPER only)
 */
export const ListUsersResponseItem = zod.object({
  id: zod.number(),
  username: zod.string(),
  name: zod.string(),
  role: zod.enum(["DEVELOPER", "DUENO", "CAJERA"]),
  isActive: zod.boolean(),
  createdAt: zod.string(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary Create a new user (DEVELOPER only)
 */
export const CreateUserBody = zod.object({
  username: zod.string(),
  name: zod.string(),
  password: zod.string(),
  role: zod.enum(["DEVELOPER", "DUENO", "CAJERA"]),
});

/**
 * @summary Update user (DEVELOPER only)
 */
export const UpdateUserParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateUserBody = zod.object({
  name: zod.string().optional(),
  isActive: zod.boolean().optional(),
  password: zod.string().optional(),
});

export const UpdateUserResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  name: zod.string(),
  role: zod.enum(["DEVELOPER", "DUENO", "CAJERA"]),
  isActive: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary List all products
 */
export const ListProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  costUsd: zod.number(),
  priceUsd: zod.number(),
  stock: zod.number(),
  minStock: zod.number(),
  createdAt: zod.string(),
});
export const ListProductsResponse = zod.array(ListProductsResponseItem);

/**
 * @summary Create a product (DUEÑO only)
 */
export const CreateProductBody = zod.object({
  name: zod.string(),
  category: zod.string(),
  costUsd: zod.number(),
  priceUsd: zod.number(),
  stock: zod.number(),
  minStock: zod.number(),
});

/**
 * @summary Update a product (DUEÑO only)
 */
export const UpdateProductParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProductBody = zod.object({
  name: zod.string(),
  category: zod.string(),
  costUsd: zod.number(),
  priceUsd: zod.number(),
  stock: zod.number(),
  minStock: zod.number(),
});

export const UpdateProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  costUsd: zod.number(),
  priceUsd: zod.number(),
  stock: zod.number(),
  minStock: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a product (DUEÑO only)
 */
export const DeleteProductParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteProductResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary List all sales
 */
export const ListSalesQueryParams = zod.object({
  from: zod.coerce.string().optional().describe("Start date (ISO)"),
  to: zod.coerce.string().optional().describe("End date (ISO)"),
});

export const ListSalesResponseItem = zod.object({
  id: zod.number(),
  cashierId: zod.number(),
  cashierName: zod.string(),
  totalUsd: zod.number(),
  totalBs: zod.number(),
  bcvRate: zod.number(),
  paymentMethod: zod.enum(["EFECTIVO_USD", "PAGO_MOVIL", "PUNTO_VENTA"]),
  items: zod.array(
    zod.object({
      productId: zod.number(),
      productName: zod.string(),
      quantity: zod.number(),
      priceUsd: zod.number(),
      costUsd: zod.number(),
    }),
  ),
  createdAt: zod.string(),
});
export const ListSalesResponse = zod.array(ListSalesResponseItem);

/**
 * @summary Register a sale
 */
export const CreateSaleBody = zod.object({
  paymentMethod: zod.enum(["EFECTIVO_USD", "PAGO_MOVIL", "PUNTO_VENTA"]),
  items: zod.array(
    zod.object({
      productId: zod.number(),
      quantity: zod.number(),
    }),
  ),
});

/**
 * @summary Get sales summary and top products (DUEÑO only)
 */
export const GetSalesSummaryQueryParams = zod.object({
  from: zod.coerce.string().optional(),
  to: zod.coerce.string().optional(),
});

export const GetSalesSummaryResponse = zod.object({
  totalSalesUsd: zod.number(),
  totalSalesBs: zod.number(),
  totalCostUsd: zod.number(),
  netProfitUsd: zod.number(),
  salesCount: zod.number(),
  topProducts: zod.array(
    zod.object({
      productId: zod.number(),
      productName: zod.string(),
      totalQuantity: zod.number(),
      totalRevenue: zod.number(),
    }),
  ),
  dailySales: zod.array(
    zod.object({
      date: zod.string(),
      totalUsd: zod.number(),
      totalBs: zod.number(),
    }),
  ),
});

/**
 * @summary Get current BCV USD/VES exchange rate
 */
export const GetBcvRateResponse = zod.object({
  rate: zod.number(),
  source: zod.enum(["SCRAPER", "MANUAL"]),
  updatedAt: zod.string(),
});

/**
 * @summary Manually set BCV rate (DEVELOPER only)
 */
export const SetBcvRateBody = zod.object({
  rate: zod.number(),
});

export const SetBcvRateResponse = zod.object({
  rate: zod.number(),
  source: zod.enum(["SCRAPER", "MANUAL"]),
  updatedAt: zod.string(),
});