# lib/api-zod/src/generated/types/dailySale.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ERP Venezuela API
 * OpenAPI spec version: 0.1.0
 */

export interface DailySale {
  date: string;
  totalUsd: number;
  totalBs: number;
}