# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ERP Venezuela API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  category: string;
  costUsd: number;
  priceUsd: number;
  stock: number;
  minStock: number;
  createdAt: string;
}